\name{calc_genoprob}
\alias{calc_genoprob}
\title{
calculate genotype probabilities
}
\description{
calculate genotype probabilities
}
\usage{
calc_genoprob(l_geno, r_geno, recombinations = c("l", "r", "total"), genotypes = c(1, 2, 3),...)
}
\arguments{
  \item{l_geno}{
left genotype
}
  \item{r_geno}{
right genotype
}
  \item{recombinations}{
recombination rates
}
  \item{genotypes}{
genotypes symbols 1,2,3
}
\item{...}{unused}
}
\author{
Lian Lian
}
\details{coming}
\examples{print("yes")}
