\name{predict_genoprob}
\alias{predict_genoprob}
\title{get the genotype probability based on Flanking markers
}
\description{coming...
}
\usage{
predict_genoprob(l_geno, r_geno, recombinations = c("l", "r", "total"),...)
}
\arguments{
  \item{l_geno}{coming..
}
  \item{r_geno}{coming..
}
  \item{recombinations}{coming...
}
\item{...}{unused}
}
\details{coming....}
\value{no value}

\author{
	Lian Lian
}

\examples{cat("predict_genoprob\n")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
