\name{recodegeno}
\alias{genoprob_IBDtoOri}
\alias{recodegeno_IBDtoOri}
\alias{recodegeno_IBD}
\alias{recodegeno}

\title{
recode genotypes
}
\description{
coming
}
\usage{
genoprob_IBDtoOri(Genoprob, parent1, parent2, origeno = c(-1, 0, 1),...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Genoprob}{
coming
}
  \item{parent1}{
coming
}
  \item{parent2}{
coming
}
  \item{origeno}{
coming
}
\item{...}{unused}
}
\details{
coming
}
\value{
coming
}
\references{
coming
}
\author{
Lian Lian
}



\examples{
	cat("recodegeno")
}
